// src/app/models/doctor.ts
export interface Doctor {
    id: number;
    doctorName: string;
    gender?: string;
    degree?: string;
    fees?: number;
    image?: string;
    mainSpecialty?: string;
    subSpecialty?: string;
    clinicName?: string;
    city?: string;
    governorate?: string;
    street?: string;
    PhoneNumber?: string;
  }
  